fix: add all

+++ src/domain/services/post.ts
@@ -1,14 +1,23 @@
 import { ICreatePostData, IUpdatePostData } from "../../application/dtos";
 import { IPostRepository } from "../../application/ports/repositories";
+import { AppError } from "../../shared";
+
 
 export class PostService {
   constructor(private repo: IPostRepository){}
 
   async create(data: ICreatePostData) {
+
     return this.repo.create(data)
   }
 
   async update(data: IUpdatePostData) {
+    const post = await this.repo.findById(data.id)
+
+    if (!post) throw new AppError("Post not found", 404);
+
+    if(post?.userId !== data.userId) throw new AppError("Forbidden", 403)
+
     return this.repo.update(data)
   }
 
@@ -20,7 +29,13 @@ export class PostService {
     return this.repo.findAll()
   }
 
-  async delete(id: string){
+  async delete(id: string, userId: string){
+        const post = await this.repo.findById(id)
+        if (!post) throw new AppError("Post not found", 404);
+
+
+    if(post?.userId !== userId) throw new AppError("Forbidden", 403)
+
     return this.repo.delete(id)
   }
 }
